+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 ### Enhancements
 
 - [TD-1667] Added a configuration option that prevents the user from sending a message when a conversation is assigned to a bot.
+- [TD-1673] Now, the conversation assignee can be updated in an existing conversation using `Kommunicate.createConversation()` method.
+
 ## [5.14.0] - 2021-03-24Z
 
 ### Enhancements